requirements.txt
asttokens==2.2.1
backcall==0.2.0
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dnspython==2.4.1
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.12.2
jedi==0.18.2
jupyter-client==8.3.0
jupyter-core==5.3.1
kaggle==1.5.16
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.4
opendatasets==0.1.22
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
text-unidecode==1.3
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
zipp==3.16.2
scikit-learn
matplotlib
seaborn
pyyaml
dill